Plus-Fix bug

+++ showroom_webapp/apps/models/loan_models.py
@@ -3,8 +3,8 @@ from django.db import models
 class Loan(models.Model):
     id = models.AutoField(primary_key=True)
     car_id = models.IntegerField()
-    loan = models.DecimalField(max_digits=12, decimal_places=2)
-    interest_rate = models.FloatField()
+    loan = models.IntegerField()
+    interest_rate = models.IntegerField()
     years = models.IntegerField(null=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
